src/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.30.5
python-dotenv==1.0.0
pydantic==2.8.2
requests==2.32.3
pdfplumber==0.11.4
Pillow==10.4.0
numpy==1.26.4
rapidfuzz==3.9.6
sentence-transformers==3.2.1
gunicorn==21.2.0
httpx==0.27.2
SQLAlchemy==2.0.36
alembic==1.13.3
